pyproject.toml
[tool.poetry]
name = "shippingservice"
version = "0.1.0"
description = ""
authors = ["Alan Bandeira <alan.p.bandeira@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
Flask-Restful = "^0.3.7"
flask-mongoengine = "^0.9.5"
pytest = "^5.2"
flask-jwt-extended = "^3.24"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
